Code improvements (#4)
* refactoring personal-info, redeem
