+++ src/pages/personal-info.js
@@ -1,9 +1,7 @@
 import Head from 'next/head'
 import React from 'react'
 import dynamic from 'next/dynamic'
-const PersonalInfoPageComponent = dynamic(() =>
-  import('@components/end-user/personal-info/PersonalInfo'),
-)
+const PersonalInfoPageComponent = dynamic(() => import('@components/end-user/personal-info'))
 
 const transition = { duration: 0.5, ease: [0.43, 0.13, 0.23, 0.96] }
 const variants = {
@@ -23,7 +21,6 @@ function PersonalInfoPage({ session }) {
 
         <meta property="og:site_name" content="Riftly"></meta>
         <meta property="keywords" content="Riftly" />
-
 
         <link rel="icon" href="/Sparkle.svg" />
       </Head>
