+++ src/pages/redeem.js
@@ -1,9 +1,7 @@
 import Head from 'next/head'
 import React from 'react'
 import dynamic from 'next/dynamic'
-const RedeemRewardsPageComponent = dynamic(() =>
-  import('@components/end-user/redeem/RedeemRewardsPage'),
-)
+const RedeemRewardsPageComponent = dynamic(() => import('@components/end-user/redeem'))
 const transition = { duration: 0.5, ease: [0.43, 0.13, 0.23, 0.96] }
 const variants = {
   hidden: { opacity: 0 },
@@ -14,7 +12,6 @@ const variants = {
 function RedeemPage({ session }) {
   return (
     <>
-
       <ChakraBox initial="hidden" animate="enter" exit="exit" variants={variants}>
         <Flex flexDirection="column" gap="16px">
           <RedeemRewardsPageComponent session={session} />
